Refactored to helper
